[rust] add benchmarks and main script

+++ 2021/src/day_1.rs
@@ -1,6 +1,6 @@
 use std::fs;
 
-pub fn part_1() -> usize {
+pub fn part_1() -> i64 {
     fs::read_to_string("day1/input")
         .expect("Unable to read input")
         .lines()
@@ -8,10 +8,10 @@ pub fn part_1() -> usize {
         .collect::<Vec<i32>>()
         .windows(2)
         .filter(|x| x[1] > x[0])
-        .count()
+        .count() as i64
 }
 
-pub fn part_2() -> usize {
+pub fn part_2() -> i64 {
     fs::read_to_string("day1/input")
         .expect("Unable to read input")
         .lines()
@@ -19,5 +19,5 @@ pub fn part_2() -> usize {
         .collect::<Vec<i32>>()
         .windows(4)
         .filter(|x| x[1] + x[2] + x[3] > x[0] + x[1] + x[2])
-        .count()
+        .count() as i64
 }
